requirements.txt
streamlit
numpy
pandas
ipykernel
tensorflow
keras
flask
pillow